[fix] useSearch 파일명 변경 폴더 이동, 로직 변경
 - hooks 폴더로 이동
 - params 핸들러를 받아오지 않도록 변경
 - isSubmitDisabled에 사용되던 useMemo 제거
